requirements.txt
requests
beautifulsoup4
pandas
openpyxl
google-generativeai
python-dotenv
playwright
httpx
